`recreateTables()` from `seedLocal()` (and move it out of main runtime module)
