+++ packages/db/src/core/consts.ts
@@ -1,9 +1,13 @@
 import { readFileSync } from 'node:fs';
+import { fileURLToPath } from 'node:url';
 
 export const PACKAGE_NAME = JSON.parse(
 	readFileSync(new URL('../../package.json', import.meta.url), 'utf8')
 ).name;
 
+export const SEED_LOCAL_IMPORT = JSON.stringify(
+	fileURLToPath(new URL('../../dist/runtime/seed-local.js', import.meta.url))
+);
 export const RUNTIME_IMPORT = JSON.stringify(`${PACKAGE_NAME}/runtime`);
 export const RUNTIME_CONFIG_IMPORT = JSON.stringify(`${PACKAGE_NAME}/runtime/config`);
