+++ packages/db/src/runtime/index.ts
@@ -1,4 +1,3 @@
-import { LibsqlError } from '@libsql/client';
 import { type ColumnBuilderBaseConfig, type ColumnDataType, sql } from 'drizzle-orm';
 import type { LibSQLDatabase } from 'drizzle-orm/libsql';
 import {
@@ -10,7 +9,6 @@ import {
 	sqliteTable,
 	text,
 } from 'drizzle-orm/sqlite-core';
-import { SEED_DEFAULT_EXPORT_ERROR, SEED_ERROR } from '../core/errors.js';
 import { type DBColumn, type DBTable } from '../core/types.js';
 import { type SerializedSQL, isSerializedSQL } from './types.js';
 
@@ -18,36 +16,6 @@ export { sql };
 export type SqliteDB = LibSQLDatabase;
 export type { Table } from './types.js';
 export { createRemoteDatabaseClient, createLocalDatabaseClient } from './db-client.js';
-
-export async function seedLocal({
-	// Glob all potential seed files to catch renames and deletions.
-	userSeedGlob,
-	integrationSeedFunctions,
-}: {
-	userSeedGlob: Record<string, { default?: () => Promise<void> }>;
-	integrationSeedFunctions: Array<() => Promise<void>>;
-}) {
-	const seedFunctions: Array<() => Promise<void>> = [];
-	const seedFilePath = Object.keys(userSeedGlob)[0];
-	if (seedFilePath) {
-		const mod = userSeedGlob[seedFilePath];
-		if (!mod.default) throw new Error(SEED_DEFAULT_EXPORT_ERROR(seedFilePath));
-		seedFunctions.push(mod.default);
-	}
-	for (const seedFn of integrationSeedFunctions) {
-		seedFunctions.push(seedFn);
-	}
-	for (const seed of seedFunctions) {
-		try {
-			await seed();
-		} catch (e) {
-			if (e instanceof LibsqlError) {
-				throw new Error(SEED_ERROR(e.message));
-			}
-			throw e;
-		}
-	}
-}
 
 export function hasPrimaryKey(column: DBColumn) {
 	return 'primaryKey' in column.schema && !!column.schema.primaryKey;
